@@ -8,15 +8,15 @@
     "jest": "27.5.1",
     "lodash": "4.17.21",
     "nanoid": "^5.0.9",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   },
   "devDependencies": {
     "@types/jest": "27.4.1",
     "@types/lodash": "4.14.179",
     "@types/node": "17.0.21",
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.13",
+    "@types/react": "^19.0.1",
+    "@types/react-dom": "^19.0.2",
     "css-loader": "6.7.0",
     "html-webpack-plugin": "5.5.0",
     "style-loader": "3.3.1",
