@@ -1,6 +1,6 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client';
 
 import App from './App';
 
-ReactDOM.render(<App/>, document.body);
+ReactDOM.createRoot(document.body).render(<App/>);
